Refactor SuumoScraper to update import path and rename property name key in response dictionary
